# internal/api/models/folder_domain.go
package models

type CreateFolderRequest struct {
	Name string `json:"name" binding:"required"`
}

type CreateFolderResponse struct {
	ID             string `json:"id"`
	OwnerID        string `json:"owner_id"`
	ParentFolderID string `json:"parent_id"`
	Name           string `json:"name"`
}

type GetFolderContentsRequest struct{}

type GetFolderContentsResponse struct {
	FolderList []*Folder `json:"folder_list"`
	FileList   []*File   `json:"file_list"`
}